requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.3
APScheduler==3.10.4
python-dotenv==1.0.1

